+++ Mods/S7_Config_de8f15f2-65a2-4ee4-a25f-7a7ab0305a58/Story/RawFiles/Lua/BootstrapClient.lua
@@ -1,0 +1,19 @@
+--  #################################################################################################################################
+--  #########                                                STATS CONFIGURATOR                                             #########
+--  #################################################################################################################################
+
+--  ===================================
+Ext.Require("S7_ConfigSettings.lua")
+Ext.Require("S7_StatsConfigurator.lua")
+--  ===================================
+
+local function S7_StatsLoader() --  Loads stats-configuration json during StatsLoaded Event.
+    local JSONstring = Ext.LoadFile("S7_Config.json") --  Only loads S7_Config.json.
+    S7_StatsConfigurator(JSONstring) --  Pass stringified JSON to StatsConfigurator()
+end
+
+--  ===============================================
+Ext.RegisterListener("StatsLoaded", S7_StatsLoader)
+--  ===============================================
+
+--  ##################################################################################################################################